Move animation into separate file
